🐛 fix: coerce those goddamn paths
