+++ packages/jest/package.json
@@ -5,7 +5,7 @@
   "publishConfig": {
     "access": "public"
   },
-  "version": "8.0.0-beta.5",
+  "version": "8.0.0-beta.6",
   "repository": {
     "type": "git",
     "url": "https://github.com/GrowflowTeam/javascript.git"
